pyproject.toml
[project]
name = "mnt-api"
version = "1.0.0"
description = "A simple API for MNT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache==0.12.2",
    "aiofiles==24.1.0",
    "aiohttp==3.9.5",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "async-googlemaps==0.0.10",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "bidict==0.23.1",
    "blurhash-python==1.2.2",
    "cachetools==5.3.3",
    "cairosvg==2.7.1",
    "certifi==2024.6.2",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "dnspython==2.6.1",
    "docstring-parser==0.16",
    "email-validator==2.1.2",
    "exceptiongroup==1.2.1",
    "exponent-server-sdk==2.1.0",
    "fal-client==0.4.0",
    "fastapi==0.111.0",
    "fastapi-cli==0.0.4",
    "fastapi-healthcheck==0.2.13",
    "frozenlist==1.4.1",
    "google-api-core>=2.24.0",
    "google-api-python-client==2.134.0",
    "google-auth==2.30.0",
    "google-auth-httplib2==0.2.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-core>=2.4.1",
    "google-cloud-logging>=3.11.0",
    "google-cloud-pubsub>=2.30.0",
    "google-cloud-resource-manager>=1.13.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-tasks>=2.14.0",
    "google-cloud-video-transcoder>=1.11.0",
    "google-crc32c==1.5.0",
    "google-genai==1.15.0",
    "google-resumable-media==2.7.1",
    "googleapis-common-protos>=1.70.0",
    "googlemaps==4.10.0",
    "google-cloud-speech==2.33.0",
    "greenlet==3.1.1",
    "grpc-google-iam-v1>=0.13.1",
    "gunicorn==20.1.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx>=0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.7",
    "jinja2==3.1.4",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "livekit==0.18.1",
    "livekit-api==0.8.0",
    "loguru==0.6.0",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "marshmallow==3.21.3",
    "mdurl==0.1.2",
    "motor==3.4.0",
    "multidict==6.0.5",
    "mypy-extensions==1.0.0",
    "numpy==1.26.4",
    "openai>=1.0.0",
    "orjson==3.10.15",
    "packaging==24.1",
    "pillow==11.1.0",
    "proto-plus>=1.26.0",
    "protobuf>=5.0,<6.0",
    "pyasn1==0.6.0",
    "pyasn1-modules==0.4.0",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings[gcp-secret-manager]==2.10.1",
    "pyee==11.1.0",
    "pygments==2.18.0",
    "pymongo==4.14.1",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "python-engineio==4.9.1",
    "python-magic==0.4.27",
    "python-multipart==0.0.9",
    "python-socketio==5.11.2",
    "pyyaml==6.0.1",
    "redis[hiredis]==5.1.1",
    "requests==2.32.3",
    "rich==13.7.1",
    "rq-scheduler==0.14.0",
    "rsa==4.9",
    "shellingham==1.5.4",
    "simple-websocket==1.0.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.5",
    "starlette==0.37.2",
    "tenacity==8.4.1",
    "tqdm==4.66.4",
    "typer==0.12.3",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "ujson==5.10.0",
    "uritemplate==4.1.1",
    "urllib3==2.2.2",
    "uvicorn==0.30.1",
    "watchfiles==1.0.4",
    "websockets==13.0",
    "wsproto==1.2.0",
    "yarl==1.9.4",
    "yt-dlp>=2025.08.22",
    "beautifulsoup4>=4.13.1",
    "json-repair>=0.42.0",
    "geoip2==5.1.0",
    "google-cloud-vision>=3.10.1",
    "langfuse==3.3.0",
    "pydantic-xml>=2.17.3",
]

[dependency-groups]
dev = ["debugpy>=1.8.14", "mypy>=1.15.0", "pre-commit>=4.1.0", "ruff>=0.9.6"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
plugins = ['pydantic.mypy']
